Forge refactor: Move Forge 'go' into separate NodeKernel.Forge module 'forge'
